+++ examples/mix.c
@@ -72,7 +72,7 @@ int main(int argc, char **argv){
     void *buffer;
     for(int i=0; i<argc-1; ++i){
       struct mp3 *mp3 = mp3s[i];
-      read = SIZE_MAX;
+      read = UINT32_MAX;
       mixed_pack_request_write(&buffer, &read, &mp3->pack);
       if(mpg123_read(mp3->handle, buffer, read, &read) != MPG123_OK){
         fprintf(stderr, "Failure during MP3 decoding: %s\n", mpg123_strerror(mp3->handle));
@@ -87,7 +87,7 @@ int main(int argc, char **argv){
       goto cleanup;
     }
     
-    size_t bytes = SIZE_MAX;
+    size_t bytes = UINT32_MAX;
     mixed_pack_request_read(&buffer, &bytes, &out->pack);
     played = out123_play(out->handle, buffer, bytes);
     if(played < bytes){
